lims/requirements.txt
# Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0
django-filter>=24.3,<25.0
django-guardian>=2.4,<3.0
django-simple-history>=3.6,<4.0
drf-spectacular>=0.27,<1.0
psycopg[binary]>=3.2,<4.0
gunicorn>=23.0,<24.0

# Async
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0

# Storage
django-storages[s3]>=1.14,<2.0
boto3>=1.35,<2.0

# Auth & Security
pyotp>=2.9,<3.0
argon2-cffi>=23.1,<24.0

# i18n (gettext是Python标准库，无需安装)

# Docs
PyYAML>=6.0,<7.0
uritemplate>=4.1,<5.0

# Dev
django-debug-toolbar>=4.4,<5.0
factory-boy>=3.3,<4.0
pytest>=8.3,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
coverage>=7.6,<8.0
bandit>=1.7,<2.0
safety>=3.2,<4.0
black>=24.10,<25.0
ruff>=0.7,<1.0
mypy>=1.13,<2.0
